requirements.txt
mcp>=1.0.0
httpx>=0.27.0
pydantic>=2.0.0
python-ulid>=2.0.0
cryptography>=42.0.0
tomli-w>=1.0.0
